requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0

# Binance API
python-binance>=1.0.29

# Environment management
python-dotenv>=1.0.0

# WebSocket support
websockets>=11.0.0

# Data visualization
plotly>=5.15.0

# Async support
aiohttp>=3.8.0
asyncio-mqtt>=0.13.0

# Data processing
scipy>=1.10.0
scikit-learn>=1.3.0

# Logging and monitoring
structlog>=23.1.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0 